api/requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
flask-cors==5.0.1